ST-81
Move particle interpolator into new project structure
Adapt tests to reflect new structure
